First commit into the master branch
Hi my name is Kevin and I am a total newb. This is my first time committing to the master branch and I don't completely understand what I am doing. Here is what has been going on. I have been developing on a branch named FitTagIOS and I feel that the app has reached a good point and wanted to save it to the master branch. The app is just about ready to enter into the testing phase and I feel that committing to the master branch at this point is a smart thing to do. There are a few known bugs and tasks that still need to be completed but that is besides the point. Here are the steps I took for this commit. I fully committed the project to FitTagIOS, switched to the master (which was empty at the time of this commit), and then downloaded a zip from FitTagIOS through the GitHub UI. I unzipped imported into the FitTagApp folder and I switched branches in order to make this commit into the master. I am almost certain there is a better way of doing this but at this moment in time i am not aware of it and as a result I am doing this hackery. Things that still need to be completed at this point in time are as follow:

Camera roll button should show most recent picture instead of plain button
Facebook & Twitter sharing needs to be done
Followers and following need to be clickable and take you to a screen that displays all of your followers/people you are following.
Font should be changed in the entire app
Hash-tagging and setting @mentions need to be highlighted throughout the app
If you leave a comment in the detail screen it needs to let the main feed view know that a change has occurred and increment by one.
Is there a way to not load the interests from parse every time? If we figure out a way to store in app or cache the array it would potentially save us some server requests.
Low network connection needs to display a screen to let the user know he is not connected to the network.
Notifications screens tableview should load more records automatically.
Push notifications for the rewards and mentions
Rewards settings needs to be completed
Searching needs to be fine tuned, the query is not very accurate
Signup Views Image needs to be saved when selected, make sure image is being saved.
UIScrollView in the map landing screen needs to scroll and be clickable depending on the notations that are shown in the map.
Users should be able to select a location from a list of locations around them. This feature will need google map and location tracking
Video should continue recording instead of restarting a new video
We shouldn't be creating a new map instance every time someone taps location. Reuse the maps tab.
Widget tools need to be tested.
There is currently a bug in the FindFriendsViewController query file, if you visit the controller it crashes. I am in the middle of finishing this view controller.